requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
cssselect2==0.7.0
Django==4.2.5
django-cors-headers==4.2.0
django-crum==0.7.9
django-easy-pdf==0.1.1
django-extensions==3.2.3
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
fonttools==4.42.1
html5lib==1.1
idna==3.4
lxml==4.9.3
oscrypto==1.3.0
Pillow==10.0.1
pycparser==2.21
pydyf==0.7.0
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
pypdf==3.16.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.8.0
tzlocal==5.0.1
uritools==4.0.2
urllib3==2.0.4
weasyprint==59.0
webencodings==0.5.1
xhtml2pdf==0.2.11
zopfli==0.2.3
